mcp_server/requirements.txt
mcp>=1.0.0
mysql-connector-python>=8.0.0
numpy>=1.21.0
psutil>=5.8.0
aiofiles>=0.8.0